requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg[binary]==3.1.17
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
alembic==1.13.1
python-multipart==0.0.6
email-validator>=2.0.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
